autoservice_crm/backend/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
jinja2>=3.1.4
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.32
alembic>=1.13.2
psycopg[binary]>=3.2.1
pydantic-settings>=2.4.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
PyJWT>=2.9.0
email-validator>=2.2.0
redis>=5.0.8
httpx>=0.27.0
aiokafka>=0.10.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp-proto-http>=1.27.0
psutil>=5.9.8
